refactor: update path for information files
