+++ src/main/java/com/techeer/f5/jmtmonster/domain/restaurant/service/MenuInfo.java
@@ -1,13 +1,17 @@
-package com.techeer.f5.jmtmonster.domain.restaurant.service.information;
+package com.techeer.f5.jmtmonster.domain.restaurant.service;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.util.List;
+
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-public class RestaurantInfo {
+public class MenuInfo {
     private Long cid;
     private String restaurantName;
+    private List menuList;
+
 }
