+++ src/main/java/com/techeer/f5/jmtmonster/domain/restaurant/service/MenuService.java
@@ -2,17 +2,12 @@ package com.techeer.f5.jmtmonster.domain.restaurant.service;
 
 import com.techeer.f5.jmtmonster.domain.restaurant.dto.mapper.RestaurantMapper;
 import com.techeer.f5.jmtmonster.domain.restaurant.dto.response.MenuInfoResponseDto;
-import com.techeer.f5.jmtmonster.domain.restaurant.dto.response.RestaurantInfoResponseDto;
-import com.techeer.f5.jmtmonster.domain.restaurant.service.information.MenuInfo;
-import com.techeer.f5.jmtmonster.domain.restaurant.service.information.RestaurantInfo;
-import net.minidev.json.parser.ParseException;
 import org.springframework.http.*;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.client.RestTemplate;
 import org.springframework.web.util.UriComponentsBuilder;
 
-import java.io.UnsupportedEncodingException;
 import java.net.URI;
 import java.nio.charset.StandardCharsets;
 import java.util.*;
