+++ src/main/java/com/techeer/f5/jmtmonster/domain/restaurant/service/RestaurantInfo.java
@@ -1,17 +1,13 @@
-package com.techeer.f5.jmtmonster.domain.restaurant.service.information;
+package com.techeer.f5.jmtmonster.domain.restaurant.service;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.util.List;
-
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-public class MenuInfo {
+public class RestaurantInfo {
     private Long cid;
     private String restaurantName;
-    private List menuList;
-
 }
